requirements.txt
APScheduler==3.7.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
six==1.15.0
tzlocal==2.1
urllib3==1.26.4
vk-api==11.9.1
